puzzle_crypto_analysis/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
requests>=2.28.0
PyJWT>=2.6.0
python-dotenv>=0.21.0
schedule>=1.1.0
pyupbit>=0.2.33
streamlit>=1.28.0
plotly>=5.17.0
ta>=0.10.2